create separate ZMAccentColor class
