+++ wire-ios-utilities/Source/Public/WireUtilities.h
@@ -24,7 +24,6 @@ FOUNDATION_EXPORT double WireUtilitiesVersionNumber;
 //! Project version string for WireUtilities.
 FOUNDATION_EXPORT const unsigned char WireUtilitiesVersionString[];
 
-
 #import <WireUtilities/NSData+ZMAdditions.h>
 #import <WireUtilities/NSData+ZMSCrypto.h>
 #import <WireUtilities/NSOperationQueue+Helpers.h>
@@ -37,6 +36,5 @@ FOUNDATION_EXPORT const unsigned char WireUtilitiesVersionString[];
 #import <WireUtilities/ZMMobileProvisionParser.h>
 #import <WireUtilities/NSString+Normalization.h>
 #import <WireUtilities/ZMEncodedNSUUIDWithTimestamp.h>
-#import <WireUtilities/ZMAccentColor.h>
 #import <WireUtilities/ZMAtomicInteger.h>
 #import <WireUtilities/ZMObjectValidationError.h>
